requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.1
pandas==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
Google-Auth-oauthlib==1.1.0
openai==1.3.3
pytest==7.4.3
httpx==0.25.2
python-jose[cryptography]==3.3.0
aiosmtplib==2.0.2
python-magic==0.4.27
Pillow==10.1.0
scikit-learn==1.3.2
xgboost==2.0.2
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2